1.setup/requirements-dev.txt
fastapi
psycopg2-binary
uvicorn
kafka-python